Refactor/Chore: gistwsv package and tests
