+++ .github/workflows/ci.yml
@@ -34,3 +34,28 @@ jobs:
       - name: Run linter
         working-directory: site
         run: yarn lint
+
+  test-gistwsv-package:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout commit
+        uses: actions/checkout@v4
+
+      - uses: pnpm/action-setup@v4
+        name: Install pnpm
+        with:
+          version: 9
+          run_install: false
+
+      - name: Use Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '18.x'
+      
+      - name: Install Dependencies
+        working-directory: packages/gist-wsv
+        run: pnpm install
+
+      - name: Run tests
+        working-directory: packages/gist-wsv
+        run: pnpm test
