+++ packages/gist-wsv/__tests__/fuzzySearch.test.ts
@@ -1,4 +1,4 @@
-import { fuzzySearch } from '../fuzzySearch';
+import { fuzzySearch } from '../src/utils/fuzzySearch';
 
 describe('fuzzySearch', () => {
   describe('Direct matching (isFuzzy=false)', () => {
