+++ packages/gist-wsv/src/components/widgets/hoverText.tsx
