+++ packages/gist-wsv/src/components/wordScaleVis/chartList.ts
