+++ packages/gist-wsv/src/components/wordScaleVis/glyphMaxMin.tsx
