+++ packages/gist-wsv/src/components/wordScaleVis/glyphText.tsx
