+++ packages/gist-wsv/src/components/wordScaleVis/hStackedBarChart.tsx
