+++ packages/gist-wsv/src/components/wordScaleVis/interface/SimpleBar.tsx
