+++ packages/gist-wsv/src/components/wordScaleVis/interface/SimpleLine.tsx
