+++ packages/gist-wsv/src/components/wordScaleVis/interface/SimpleMaxMin.tsx
