+++ packages/gist-wsv/src/components/wordScaleVis/interface/SimpleStackedBar.tsx
