+++ packages/gist-wsv/src/components/wordScaleVis/interface/index.ts
