+++ packages/gist-wsv/src/components/wordScaleVis/interface/types.ts
