+++ packages/gist-wsv/src/components/wordScaleVis/vBarChart.tsx
