+++ packages/gist-wsv/src/renderer/extremeTextRenderer.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
-import { EntitySpec, GistvisSpec } from '../types';
+import { EntitySpec, GistvisSpec } from '../components/types';
 import * as d3 from 'd3';
-import HoverText from '../widgets/hoverText';
-import { GlyphMaxMin } from '../wordScaleVis/chartList';
+import HoverText from '../components/widgets/hoverText';
+import { GlyphMaxMin } from '../components/wordScaleVis/chartList';
 import { getHighlightPos, getProductionVisSpec, getUniqueEntities } from '../utils/postProcess';
 import useTrackVisit from '../utils/useTrack';
