+++ packages/gist-wsv/src/renderer/trendTextRenderer.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
-import { DataPoint, EntitySpec, GistvisSpec, TrendAttribute, TrendOptions } from '../types';
+import { DataPoint, EntitySpec, GistvisSpec, TrendAttribute, TrendOptions } from '../components/types';
 import * as d3 from 'd3';
-import HoverText from '../widgets/hoverText';
-import { LineChart } from '../wordScaleVis/chartList';
+import HoverText from '../components/widgets/hoverText';
+import { LineChart } from '../components/wordScaleVis/chartList';
 import { getHighlightPos, getProductionVisSpec, getUniqueEntities } from '../utils/postProcess';
 import useTrackVisit from '../utils/useTrack';
