--- packages/gist-wsv/src/style/page.css
@@ -1,147 +1,0 @@
-.article {
-  position: absolute;
-  top: 65%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  overflow: auto;
-  min-height: 600px;
-}
-
-.text {
-  font-size: medium;
-  font-style: bold;
-  font-weight: bold;
-}
-
-.pos_text {
-  font-size: medium;
-  font-style: normal;
-  color: crimson;
-  font-weight: bold;
-  font-size: 2ch;
-  margin-right: 5px;
-}
-
-.neg_text {
-  font-size: medium;
-  font-style: normal;
-  color: limegreen;
-  font-weight: bold;
-  font-size: 2ch;
-  margin-right: 5px;
-}
-
-.r_text {
-  font-size: medium;
-  font-style: normal;
-  font-weight: bold;
-  margin-right: 5px;
-}
-
-.spec_value {
-  font-family: sans-serif;
-  font-size: 2ch;
-  font-style: normal;
-  font-weight: 550;
-}
-
-.ano_value {
-  font-family: sans-serif;
-  font-size: normal;
-  font-style: normal;
-  color: red;
-}
-
-.max_value {
-  font-family: sans-serif;
-  font-size: normal;
-  font-style: normal;
-  color: red;
-  font-weight: bold;
-}
-
-.min_value {
-  font-family: sans-serif;
-  font-size: normal;
-  font-style: normal;
-  color: green;
-  font-weight: bold;
-}
-
-.title {
-  font-size: 80px;
-  font: bolder;
-  position: absolute;
-  top: 10%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.input_container {
-  position: absolute;
-  top: 20%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.input {
-  width: 500px;
-  height: 30px;
-}
-
-.input_button {
-  width: 100px;
-  height: 35px;
-  position: relative;
-  left: 2px;
-  display: inline-block;
-  font-size: 16px;
-  color: #fff;
-  background-color: #4b0082;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  text-align: center;
-  text-decoration: none;
-  font-family: Georgia;
-  transition: background-color 0.3s;
-}
-
-.pre-wrap {
-  white-space: pre-wrap;
-}
-/* span {
-  line-height: 30px; 
-  white-space: pre-wrap;
-  font-weight: bold;
-} */
-
-.ant-carousel .slick-prev,
-.ant-carousel .slick-next,
-.ant-carousel .slick-prev:hover,
-.ant-carousel .slick-next:hover,
-.ant-carousel .slick-prev:focus,
-.ant-carousel .slick-next:focus {
-  font-size: inherit;
-  color: rgba(76, 144, 226, 0.3) !important;
-}
-
-.ant-carousel .slick-prev::before {
-  color: rgba(76, 144, 226, 0.3) !important;
-}
-
-.slick-arrow.slick-prev {
-  font-size: 20px !important;
-}
-
-.ant-carousel .slick-next::before {
-  color: rgba(76, 144, 226, 0.3) !important;
-}
-
-.slick-arrow.slick-next {
-  font-size: 20px !important;
-}
-
-.hide-scrollbar::-webkit-scrollbar {
-  display: none;
-}
