+++ packages/gist-wsv/tsconfig.app.json
@@ -18,6 +18,6 @@
     "allowJs": true,
     "esModuleInterop": true
   },
-  "include": ["src"],
+  "include": ["src", "__tests__"],
   "exclude": ["node_modules", "dist"]
 }
